backend/requirements_workers.txt
fastapi
uvicorn>=0.15.0
sqlalchemy
psycopg2-binary
minio
pymilvus
redis
requests
dramatiq[redis]
python-dotenv
python-docx
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-jose
python-multipart
puremagic
cryptography
openai
beautifulsoup4
lxml
PyMuPDF
olefile
jinja2
transitions
zipstream
psutil
Pillow
click>=8.1.7
mineru[core]